[REFACT] 작가가 자신의 예약 시간 형식을 등록, 수정, 삭제, 매핑하는 로직 Reservation클래스에서 분리
